Created a monster with user input.
Created a monster based on user input. Practiced using variables and
Scanner methods.

+++ src/marshmallow/model/MarshmallowMonster.java
@@ -72,12 +72,12 @@ public class MarshmallowMonster
 		this.monsterNoses = monsterNoses;
 	}
 	
-	public void setMonsterHair(int monsterHair)
+	public void setMonsterHair(double monsterHair)
 	{
 		this.monsterHair = monsterHair;
 	}
 	
-	public void setMonsterLegs(int monsterLegs)
+	public void setMonsterLegs(double monsterLegs)
 	{
 		this.monsterLegs = monsterLegs;
 	}
